requirements.txt
Flask==3.0.3
requests==2.32.3
PyYAML==6.0.2
bleach==6.2.0
